refactor: move useListbox to its own folder.
