+++ packages/mui-base/src/MenuUnstyled/useMenu.types.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { UseListboxRootSlotProps } from '../ListboxUnstyled';
+import { UseListboxRootSlotProps } from '@mui/base/useListbox';
 
 export interface MenuItemMetadata {
   id: string;
