+++ packages/mui-base/src/useListbox/defaultListboxReducer.ts
