+++ packages/mui-base/src/useListbox/index.ts
@@ -1,4 +1,7 @@
 export { default as useListbox } from './useListbox';
+export { default } from './useListbox';
+export * from './useListbox';
+
 export { default as defaultListboxReducer } from './defaultListboxReducer';
 
 export * from './useListbox.types';
